--- .idea/inspectionProfiles/Project_Default.xml
@@ -1,19 +1,0 @@
-<component name="InspectionProjectProfileManager">
-  <profile version="1.0">
-    <option name="myName" value="Project Default" />
-    <inspection_tool class="DuplicatedCode" enabled="true" level="WEAK WARNING" enabled_by_default="true">
-      <Languages>
-        <language minSize="60" name="Python" />
-      </Languages>
-    </inspection_tool>
-    <inspection_tool class="PyUnresolvedReferencesInspection" enabled="true" level="WARNING" enabled_by_default="true">
-      <option name="ignoredIdentifiers">
-        <list>
-          <option value="str.b64decode" />
-          <option value="super._retry" />
-          <option value="tuple.append" />
-        </list>
-      </option>
-    </inspection_tool>
-  </profile>
-</component>
