+++ src/backend/core/Item.py
@@ -1,5 +1,3 @@
-# items.py
-
 class Item:
     instances = []
     def __init__(self, nome, preco_mercado, preco_craft, cidade, encantamento):
@@ -9,17 +7,3 @@ class Item:
         self.cidade = cidade
         self.encantamento = encantamento
         Item.instances.append(self)
-
-
-class CraftingSlot:
-    instances = []
-    # [Lista com todas as recipies, temporaria até nao termos PostGres]
-
-    def __init__(self, craft_id, item_destino_nome, qtd, item_source_nome):
-        self.craft_id = craft_id
-        self.item_destino_nome = item_destino_nome
-        self.qtd = qtd
-        self.item_source_nome = item_source_nome
-        CraftingSlot.instances.append(self)
-
-
